Move Convert processor to convert module (apply review comment)
