+++ pilkit/processors/__init__.py
@@ -12,6 +12,6 @@ from the filesystem.
 
 from .base import *
 from .crop import *
-from .image import *
+from .convert import *
 from .overlay import *
 from .resize import *
